Fast_Api/requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ORM and Database
sqlalchemy==1.4.52
pymysql==1.1.0

# JWT Authentication
fastapi-jwt-auth==0.5.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# Pydantic compatible with fastapi-jwt-auth
pydantic==1.10.15

# Response & HTTP tools
python-multipart==0.0.9
requests==2.31.0

# Optional Dev Tools
python-dotenv==1.0.1

# Password hashing (used in werkzeug.security)
werkzeug==3.0.1
email-validator==1.3.1
